requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cleanup==8.0.0
django-import-export==3.2.0
django-jazzmin==2.6.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
python-dotenv==1.0.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
